debug vec type

+++ lab/case-3/dataset.h
@@ -23,7 +23,7 @@
 // number of incr operations on every cache line
 #define NUM_OP 16
 
-typedef intptr_t vec_t[LEN_VEC];
+typedef int vec_t[LEN_VEC];
 
 /* Force test_vec to be allocated in .data */
 static vec_t test_vec = { 1 };
